code

+++ lib/firebase/users_calls.dart
@@ -3,7 +3,6 @@ import 'dart:io';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_storage/firebase_storage.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:pharmabox/model/user_models/non_titulaire.dart';
 
 class UserService {
@@ -13,29 +12,31 @@ class UserService {
   Future creerNonTitulaire(NonTitulaire user) async {
     User? authUser = firebaseAuth.currentUser!;
     print(user.photoUrl);
-    if(user.photoUrl!= ''){
-  final storageRef = FirebaseStorage.instance.ref();
-    final profileRef = storageRef
-        .child('profile_picture')
-        .child(authUser.uid)
-        .child('${authUser.uid}.jpg');
-    await profileRef.getDownloadURL().then((value) async {
-      await profileRef.delete();
-      await profileRef.putFile(File(user.photoUrl));
-      await profileRef.getDownloadURL().then((value) async {
-        print("first error");
-        user.photoUrl = value;
-      });
-    }).catchError((error) async {
-      print("second error");
+    if (user.photoUrl != '') {
+      File file = File(user.photoUrl);
+      if (file.existsSync()) {
+        final storageRef = FirebaseStorage.instance.ref();
+        final profileRef = storageRef
+            .child('profile_picture')
+            .child(authUser.uid)
+            .child('${authUser.uid}.jpg');
+        try {
+          print(authUser.uid);
+          await profileRef.getDownloadURL();
+          await profileRef.delete();
+          await profileRef.putFile(file);
+          user.photoUrl = await profileRef.getDownloadURL();
+        } catch (e) {
+          print(e.toString());
 
-      await profileRef.putFile(File(user.photoUrl)).then(
-          (p0) async => await profileRef.getDownloadURL().then((value) async {
-                user.photoUrl = value;
-              }));
-    });
+          if (e.toString() ==
+              "[firebase_storage/object-not-found] No object exists at the desired reference.") {
+            await profileRef.putFile(file);
+            user.photoUrl = await profileRef.getDownloadURL();
+          }
+        }
+      }
     }
-   
 
     await _firebaseFirestore
         .collection("users")
